reversed logic in getNativeBindings to reduce indentation
Reading this code again I found a way to reduces the complexity by doing
the various checks incrementally and always returning the stub if any
check fails.  That way, we don't need nested conditionals which always
tend to be harder to read and understand than a sequence of checks like
it is now.